+++ bin/check-snmp.rb
@@ -46,6 +46,12 @@ class CheckSNMP < Sensu::Plugin::Check::CLI
          short: '-c critical',
          default: '20'
 
+  option :reverse,
+         short: '-r',
+         description: 'Revese the logic of the test for warning and critical thresholds',
+         boolean: true,
+         default: false
+
   option :match,
          short: '-m match',
          description: 'Regex pattern to match against returned value'
@@ -94,7 +100,11 @@ class CheckSNMP < Sensu::Plugin::Check::CLI
     rescue => e
       unknown "An unknown error occured: #{e.inspect}"
     end
-    operators = { 'le' => :<=, 'ge' => :>= }
+    operators = if config[:reverse]
+                  { 'le' => :>=, 'ge' => :<= }
+                else
+                  { 'le' => :<=, 'ge' => :>= }
+                end
     symbol = operators[config[:comparison]]
 
     response.each_varbind do |vb|